+++ TrinketTinker/Extras/GameItemQuery.cs
@@ -44,7 +44,7 @@ public static class GameItemQuery
     /// <param name="context"></param>
     /// <param name="value"></param>
     /// <returns></returns>
-    private static bool TryGetOptionalOrRandomOrPlus(
+    private static bool TryGetOptionalOrRandom(
         string[] array,
         int index,
         int maxValue,
@@ -59,9 +59,6 @@ public static class GameItemQuery
                 case RANDOM:
                     value = (context.Random ?? Random.Shared).Next(maxValue);
                     return true;
-                case ANY:
-                    value = -1;
-                    return false;
                 default:
                     break;
             }
@@ -133,9 +130,9 @@ public static class GameItemQuery
         {
             if (trinket.GetEffect() is TrinketTinkerEffect effect)
             {
-                if (TryGetOptionalOrRandomOrPlus(array, 1, effect.MaxLevel, context, out int level))
+                if (TryGetOptionalOrRandom(array, 1, effect.MaxLevel, context, out int level))
                     effect.SetLevel(trinket, level);
-                if (TryGetOptionalOrRandomOrPlus(array, 2, effect.MaxVariant, context, out int variant))
+                if (TryGetOptionalOrRandom(array, 2, effect.MaxVariant, context, out int variant))
                     effect.SetVariant(trinket, variant);
             }
             return [new ItemQueryResult(trinket)];
@@ -186,7 +183,7 @@ public static class GameItemQuery
         if (effect.MaxVariant == 0)
             return ItemQueryResolver.Helpers.ErrorResult(key, arguments, logError, $"Trinket has no variants.");
 
-        TryGetOptionalOrRandomOrPlus(array, 1, effect.MaxLevel, context, out int level);
+        TryGetOptionalOrRandom(array, 1, effect.MaxLevel, context, out int level);
 
         List<ItemQueryResult> createdTrinkets = [];
         for (int variant = 0; variant < effect.MaxVariant; variant++)